+++ docs/changelog.rst
@@ -44,6 +44,7 @@ Here we list a changelog of pybv.
 Changelog
 ~~~~~~~~~
 - Updated ``_chk_events`` to accept label values >= 0, by `Pierre Cutellic`_: (:gh:`95`)
+- Updated ``_chk_events`` to accept duration == 0, by `Clemens Brunner`_: (:gh:`96`)
 
 0.7.3 (2022-06-04)
 ==================